Adjusted testing structure
