+++ includes/test-arrays.h
@@ -1,0 +1,4 @@
+#ifndef TEST_ARRAYS_H
+#define TEST_ARRAYS_H
+    void run_array_tests(void);
+#endif
